refactor: Flatten struct, use moved value
